update: fix cpplint warnings and moved header files

+++ src/main.c
@@ -1,10 +1,12 @@
-#include "../include/logger.h"
-#include "args_parser.h"
-#include "error/exit_codes.h"
-#include "dirs.h"
-#include "file_status.h"
 #include <stdint.h>
 #include <stdio.h>
+#include <errno.h>
+
+#include "../include/args_parser.h"
+#include "../include/dirs.h"
+#include "../include/error/exit_codes.h"
+#include "../include/file_status.h"
+#include "../include/logger.h"
 
 int main(int argc, char* argv[], char* envp[]) {
     // cmd_args_t args;
@@ -28,20 +30,21 @@ int main(int argc, char* argv[], char* envp[]) {
         default:
             return err;
     }
-    
+
     // struct stat status;
     // get_status(argv[1], &status);
     // if (is_dir(&status)) recursive_change_mod(argv[1]);
-    
+
     cmd_args_t args;
     parse_args(&args, argc, argv, envp);
-    printf("Args obtained:\nR - %d\nv  - %d\nc  - %d\n", args.options.recursive, args.options.verbose, args.options.verbose_on_modify);
-    printf("%o, %d, %d, %d\n", args.mode.permission_octal, args.mode.permission_types.type_u, args.mode.permission_types.type_g, args.mode.permission_types.type_o);
+    printf("Args obtained:\nR - %d\nv  - %d\nc  - %d\n", args.options.recursive,
+           args.options.verbose, args.options.verbose_on_modify);
+    printf("%o, %d, %d, %d\n", args.mode.permission_octal,
+           args.mode.permission_types.type_u, args.mode.permission_types.type_g,
+           args.mode.permission_types.type_o);
 
-    printf("\nFiles: %d %d %d\n",
-           (int)(args.files_end - args.files_start),
-           (int)(args.files_start),
-           (int)(args.files_end));
+    printf("\nFiles: %lu %zu %zu\n", (args.files_end - args.files_start),
+           (args.files_start), (args.files_end));
     for (size_t i = args.files_start; i < args.files_end; i++) {
         printf("- %s\n", argv[i]);
     }
@@ -49,6 +52,6 @@ int main(int argc, char* argv[], char* envp[]) {
     if (close_log()) {
         return errno;
     }
-    
+
     return 0;
 }
